Google_calendar_API/Ksana_Project/requirements.txt
apple-calendar-integration==0.0.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
datefinder==0.7.0
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
oauthlib==3.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.0
pytz==2019.3
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.6
virtualenv==16.7.6
